// frontend/src/api/stub/models/Table.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TableStatus } from './TableStatus';
import {
    TableStatusFromJSON,
    TableStatusFromJSONTyped,
    TableStatusToJSON,
    TableStatusToJSONTyped,
} from './TableStatus';

/**
 * 
 * @export
 * @interface Table
 */
export interface Table {
    /**
     * 
     * @type {string}
     * @memberof Table
     */
    tableName: string;
    /**
     * 
     * @type {number}
     * @memberof Table
     */
    capacity: number;
    /**
     * 
     * @type {TableStatus}
     * @memberof Table
     */
    status: TableStatus;
    /**
     * 
     * @type {number}
     * @memberof Table
     */
    id?: number | null;
}



/**
 * Check if a given object implements the Table interface.
 */
export function instanceOfTable(value: object): value is Table {
    if (!('tableName' in value) || value['tableName'] === undefined) return false;
    if (!('capacity' in value) || value['capacity'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function TableFromJSON(json: any): Table {
    return TableFromJSONTyped(json, false);
}

export function TableFromJSONTyped(json: any, ignoreDiscriminator: boolean): Table {
    if (json == null) {
        return json;
    }
    return {
        
        'tableName': json['table_name'],
        'capacity': json['capacity'],
        'status': TableStatusFromJSON(json['status']),
        'id': json['id'] == null ? undefined : json['id'],
    };
}

export function TableToJSON(json: any): Table {
    return TableToJSONTyped(json, false);
}

export function TableToJSONTyped(value?: Table | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'table_name': value['tableName'],
        'capacity': value['capacity'],
        'status': TableStatusToJSON(value['status']),
        'id': value['id'],
    };
}