// frontend/src/api/stub/models/TableBase.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TableStatus } from './TableStatus';
import {
    TableStatusFromJSON,
    TableStatusFromJSONTyped,
    TableStatusToJSON,
    TableStatusToJSONTyped,
} from './TableStatus';

/**
 * 
 * @export
 * @interface TableBase
 */
export interface TableBase {
    /**
     * 
     * @type {string}
     * @memberof TableBase
     */
    tableName: string;
    /**
     * 
     * @type {number}
     * @memberof TableBase
     */
    capacity: number;
    /**
     * 
     * @type {TableStatus}
     * @memberof TableBase
     */
    status: TableStatus;
}



/**
 * Check if a given object implements the TableBase interface.
 */
export function instanceOfTableBase(value: object): value is TableBase {
    if (!('tableName' in value) || value['tableName'] === undefined) return false;
    if (!('capacity' in value) || value['capacity'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function TableBaseFromJSON(json: any): TableBase {
    return TableBaseFromJSONTyped(json, false);
}

export function TableBaseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TableBase {
    if (json == null) {
        return json;
    }
    return {
        
        'tableName': json['table_name'],
        'capacity': json['capacity'],
        'status': TableStatusFromJSON(json['status']),
    };
}

export function TableBaseToJSON(json: any): TableBase {
    return TableBaseToJSONTyped(json, false);
}

export function TableBaseToJSONTyped(value?: TableBase | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'table_name': value['tableName'],
        'capacity': value['capacity'],
        'status': TableStatusToJSON(value['status']),
    };
}