// frontend/src/api/stub/models/TableUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TableStatus } from './TableStatus';
import {
    TableStatusFromJSON,
    TableStatusFromJSONTyped,
    TableStatusToJSON,
    TableStatusToJSONTyped,
} from './TableStatus';

/**
 * 
 * @export
 * @interface TableUpdate
 */
export interface TableUpdate {
    /**
     * 
     * @type {string}
     * @memberof TableUpdate
     */
    tableName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof TableUpdate
     */
    capacity?: number | null;
    /**
     * 
     * @type {TableStatus}
     * @memberof TableUpdate
     */
    status?: TableStatus | null;
}



/**
 * Check if a given object implements the TableUpdate interface.
 */
export function instanceOfTableUpdate(value: object): value is TableUpdate {
    return true;
}

export function TableUpdateFromJSON(json: any): TableUpdate {
    return TableUpdateFromJSONTyped(json, false);
}

export function TableUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TableUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'tableName': json['table_name'] == null ? undefined : json['table_name'],
        'capacity': json['capacity'] == null ? undefined : json['capacity'],
        'status': json['status'] == null ? undefined : TableStatusFromJSON(json['status']),
    };
}

export function TableUpdateToJSON(json: any): TableUpdate {
    return TableUpdateToJSONTyped(json, false);
}

export function TableUpdateToJSONTyped(value?: TableUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'table_name': value['tableName'],
        'capacity': value['capacity'],
        'status': TableStatusToJSON(value['status']),
    };
}