// frontend/src/api/stub/models/TableStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TableStatus = {
    Available: 'available',
    Reserved: 'reserved',
    Occupied: 'occupied'
} as const;
export type TableStatus = typeof TableStatus[keyof typeof TableStatus];


export function instanceOfTableStatus(value: any): boolean {
    for (const key in TableStatus) {
        if (Object.prototype.hasOwnProperty.call(TableStatus, key)) {
            if (TableStatus[key as keyof typeof TableStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TableStatusFromJSON(json: any): TableStatus {
    return TableStatusFromJSONTyped(json, false);
}

export function TableStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): TableStatus {
    return json as TableStatus;
}

export function TableStatusToJSON(value?: TableStatus | null): any {
    return value as any;
}

export function TableStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): TableStatus {
    return value as TableStatus;
}